pyproject.toml
[project]
name = "isvicre-cakisi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.3",
    "filetype>=1.2.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "markdown>=3.10",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=12.0.0",
    "puremagic>=1.30",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.3.0",
    "python-multipart>=0.0.20",
    "qrcode[pil]>=8.2",
    "uvicorn>=0.38.0",
]
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"
[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501", "W291", "W293"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
